feat: create LessonsList component
