+++ src/constants/other.ts
@@ -1,0 +1,1 @@
+export const LOCKED = "locked";
