+++ src/pages/courseDetailsPage/CourseDetailsPage.styled.ts
@@ -1,12 +1,6 @@
 import { NavLink } from "react-router-dom";
 import styled from "styled-components";
 import { MdArrowBackIos } from "react-icons/md";
-import { FaLock } from "react-icons/fa";
-import theme from "styles/theme";
-
-interface LessonTitleProps extends React.HTMLProps<HTMLHeadingElement> {
-  active?: boolean;
-}
 
 export const ArrowIcon = styled(MdArrowBackIos)`
   width: 20px;
@@ -30,12 +24,6 @@ export const CourseTitle = styled.h1`
   text-align: center;
 `;
 
-export const Title = styled.p`
-  margin-bottom: 15px;
-  font-weight: 700;
-  font-size: 32px;
-`;
-
 export const CourseDescription = styled.p`
   margin-bottom: 16px;
   text-align: center;
@@ -44,8 +32,6 @@ export const CourseDescription = styled.p`
     margin-bottom: 20px;
   }
 `;
-
-export const CourseLessonsList = styled.ol``;
 
 export const BackLink = styled(NavLink)`
   position: absolute;
@@ -71,43 +57,4 @@ export const BackLink = styled(NavLink)`
   }
 `;
 
-export const LockIcon = styled(FaLock)``;
-
-export const TitleBox = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const LessonTitle = styled.h2<LessonTitleProps>`
-  margin-right: 10px;
-  font-weight: 500;
-  color: ${(p) => (p.active ? theme.colors.title : "inherit")};
-  transition: transform 250ms cubic-bezier(0.4, 0, 0.2, 1);
-
-  &:hover {
-    transform: scale(1.03);
-  }
-`;
-
 export const LessonLink = styled.a``;
-
-export const CurrentLessContainer = styled.div`
-  margin-bottom: 20px;
-`;
-
-export const CurrentLesson = styled.p`
-  margin: 8px 0;
-  font-size: 20px;
-  font-weight: 600;
-`;
-
-export const CurrentText = styled.span`
-  font-weight: 400;
-`;
-
-export const CourseLesson = styled.li`
-  &:not(:last-child) {
-    margin-bottom: 12px;
-  }
-  cursor: pointer;
-`;
